\name{simpack-package}
\alias{simpack-package}
\alias{simpack}
\docType{package}
\title{
\packageTitle{simpack}
}
\description{
\packageDescription{simpack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{simpack}
\packageIndices{simpack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{simpack}

Maintainer: \packageMaintainer{simpack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
